requirements/base.txt
Django==5.1.4
djangorestframework==3.15.1
mysqlclient==2.2.7
mypy_extensions==1.1.0
python-dotenv==1.0.1
